[#1347] Clean up the code
### What's done:

 * `IndentationError` moved to a separate source file.
 * Top-level extensions moved from `IndentationRule.kt` to `StringUtils.kt`.
